requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile
#
-e git+https://github.com/LCBRU/lbrc_flask.git@main#egg=lbrc_flask
    # via -r requirements.in
alembic==1.11.1
    # via -r requirements.in
amqp==5.1.1
    # via kombu
arrow==1.2.3
    # via lbrc-flask
async-timeout==4.0.2
    # via redis
attrs==22.2.0
    # via jsonschema
bcrypt==4.0.1
    # via
    #   flask-security-too
    #   lbrc-flask
billiard==3.6.4.0
    # via celery
bleach==6.0.0
    # via flask-security-too
blinker==1.6.2
    # via
    #   flask
    #   flask-mail
    #   flask-principal
cachetools==5.3.1
    # via -r requirements.in
cairocffi==1.5.0
    # via
    #   cairosvg
    #   weasyprint
cairosvg==2.7.0
    # via weasyprint
celery[redis]==5.2.7
    # via -r requirements.in
certifi==2022.12.7
    # via requests
cffi==1.15.1
    # via
    #   cairocffi
    #   weasyprint
chardet==5.1.0
    # via lbrc-flask
charset-normalizer==3.1.0
    # via requests
click==8.1.3
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
    #   flask
click-didyoumean==0.3.0
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.2.0
    # via celery
cssselect2==0.7.0
    # via
    #   cairosvg
    #   weasyprint
defusedxml==0.7.1
    # via cairosvg
dnspython==2.3.0
    # via email-validator
email-validator==1.3.1
    # via
    #   flask-security-too
    #   lbrc-flask
et-xmlfile==1.1.0
    # via openpyxl
flask==2.3.2
    # via
    #   -r requirements.in
    #   flask-admin
    #   flask-api
    #   flask-login
    #   flask-mail
    #   flask-mailman
    #   flask-principal
    #   flask-security-too
    #   flask-sqlalchemy
    #   flask-weasyprint
    #   flask-wtf
    #   lbrc-flask
flask-admin==1.6.1
    # via lbrc-flask
flask-api==3.0.post1
    # via lbrc-flask
flask-login==0.6.2
    # via flask-security-too
flask-mail==0.9.1
    # via lbrc-flask
flask-mailman==0.3.0
    # via flask-security-too
flask-principal==0.4.0
    # via flask-security-too
flask-security-too[common,fsqla]==5.2.0
    # via -r requirements.in
flask-sqlalchemy==3.0.3
    # via
    #   flask-security-too
    #   lbrc-flask
flask-weasyprint==1.0.0
    # via
    #   -r requirements.in
    #   lbrc-flask
flask-wtf==1.1.1
    # via flask-security-too
greenlet==2.0.2
    # via sqlalchemy
html5lib==1.1
    # via weasyprint
idna==3.4
    # via
    #   email-validator
    #   requests
importlib-resources==5.12.0
    # via flask-security-too
itsdangerous==2.1.2
    # via
    #   flask
    #   flask-wtf
jinja2==3.1.2
    # via flask
jsonschema==4.17.3
    # via lbrc-flask
kombu==5.2.4
    # via celery
mako==1.2.4
    # via alembic
markdown==3.4.1
    # via lbrc-flask
markdown-it-py==2.2.0
    # via rich
markupsafe==2.1.2
    # via
    #   jinja2
    #   mako
    #   werkzeug
    #   wtforms
mdurl==0.1.2
    # via markdown-it-py
mkdocs-material-extensions==1.1.1
    # via flask-mailman
openpyxl==3.1.2
    # via lbrc-flask
passlib==1.7.4
    # via flask-security-too
pillow==9.4.0
    # via
    #   cairosvg
    #   weasyprint
prompt-toolkit==3.0.38
    # via click-repl
pyasn1==0.4.8
    # via
    #   pyasn1-modules
    #   python-ldap
pyasn1-modules==0.2.8
    # via python-ldap
pycparser==2.21
    # via cffi
pygal==3.0.0
    # via lbrc-flask
pygments==2.14.0
    # via rich
pymysql==1.0.2
    # via -r requirements.in
pyphen==0.14.0
    # via weasyprint
pyrsistent==0.19.3
    # via jsonschema
python-dateutil==2.8.2
    # via arrow
python-dotenv==1.0.0
    # via
    #   -r requirements.in
    #   lbrc-flask
python-ldap==3.4.3
    # via lbrc-flask
pytz==2022.7.1
    # via
    #   celery
    #   lbrc-flask
redis==4.5.2
    # via celery
requests==2.28.2
    # via -r requirements.in
rich==13.3.2
    # via lbrc-flask
six==1.16.0
    # via
    #   bleach
    #   click-repl
    #   html5lib
    #   python-dateutil
sqlalchemy==2.0.7
    # via
    #   -r requirements.in
    #   alembic
    #   flask-security-too
    #   flask-sqlalchemy
    #   sqlalchemy-utils
sqlalchemy-utils==0.41.1
    # via flask-security-too
tinycss2==1.2.1
    # via
    #   cairosvg
    #   cssselect2
    #   weasyprint
typing-extensions==4.5.0
    # via
    #   alembic
    #   sqlalchemy
urllib3==1.26.15
    # via requests
vine==5.0.0
    # via
    #   amqp
    #   celery
    #   kombu
wcwidth==0.2.6
    # via prompt-toolkit
weasyprint==52.5
    # via
    #   -r requirements.in
    #   flask-weasyprint
webencodings==0.5.1
    # via
    #   bleach
    #   cssselect2
    #   html5lib
    #   tinycss2
werkzeug==2.3.4
    # via
    #   flask
    #   flask-login
wtforms==3.0.1
    # via
    #   flask-admin
    #   flask-security-too
    #   flask-wtf
xlrd==2.0.1
    # via lbrc-flask

# The following packages are considered to be unsafe in a requirements file:
# setuptools
